+++ src/main/java/ladysnake/blast/client/BlastClient.java
@@ -49,7 +49,13 @@ public class BlastClient implements ClientModInitializer {
                 BlastEntities.DIRT_BOMB,
                 BlastEntities.DIRT_TRIGGER_BOMB,
                 BlastEntities.PEARL_BOMB,
-                BlastEntities.PEARL_TRIGGER_BOMB
+                BlastEntities.PEARL_TRIGGER_BOMB,
+                BlastEntities.AMETHYST_BOMB,
+                BlastEntities.AMETHYST_TRIGGER_BOMB,
+                BlastEntities.FROST_BOMB,
+                BlastEntities.FROST_TRIGGER_BOMB,
+                BlastEntities.SLIME_BOMB,
+                BlastEntities.SLIME_TRIGGER_BOMB
         );
         registerBlockEntityRender(BlastEntities.GUNPOWDER_BLOCK, e -> BlastBlocks.GUNPOWDER_BLOCK.getDefaultState());
         registerBlockEntityRender(BlastEntities.STRIPMINER, StripminerEntity::getState);