+++ src/main/java/ladysnake/blast/common/entity/FrostBombEntity.java
@@ -1,0 +1,31 @@
+package ladysnake.blast.common.entity;
+
+import ladysnake.blast.common.init.BlastEntities;
+import ladysnake.blast.common.init.BlastItems;
+import ladysnake.blast.common.world.CustomExplosion;
+import ladysnake.blast.common.world.EntityExplosion;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.item.Item;
+import net.minecraft.world.World;
+import net.minecraft.world.explosion.Explosion;
+
+public class FrostBombEntity extends BombEntity {
+    public FrostBombEntity(EntityType<? extends BombEntity> entityType, World world) {
+        super(entityType, world);
+    }
+
+    public FrostBombEntity(EntityType<? extends BombEntity> entityType, World world, LivingEntity livingEntity) {
+        super(entityType, world, livingEntity);
+    }
+
+    @Override
+    protected Item getDefaultItem() {
+        return BlastItems.FROST_BOMB;
+    }
+
+    @Override
+    protected CustomExplosion getExplosion() {
+        return new EntityExplosion(this.world, this.getOwner(), this.getX(), this.getY(), this.getZ(), BlastEntities.ICICLE, 250, 0.6f);
+    }
+}